[NOT DONE]: Drive to Algae
Added a method (intakeAlgae -> void) that tries to drive to the algae (with overshoot) so that it can be intaked.

The function has a lot of TODO comments that need to be fixed or thought about.

The function also does not intake the algae yet, which should be changed.

The function name may change, hence it is experimental.